# Speed up `Lint/Debugger` by 45%.