Using a set is a lot more efficient than a list.

Rubocop takes up a decent chunk of time in our CI build, so I wanted to see if there's anything I could do to speed it up. I recorded the timings and ran this over rubocop's own codebase. This is a small win, though anyone is welcome to take a crack at the other cops.

Before
````
Slowest Cops
2.501s: Style/ClosingParenthesisIndentation
2.315s: Style/MultilineOperationIndentation
1.197s: Style/IndentationWidth
1.103s: Lint/BlockAlignment
1.010s: Style/FirstParameterIndentation
0.797s: Lint/DeprecatedClassMethods
0.710s: Metrics/LineLength
0.683s: Lint/Debugger
0.618s: Style/SignalException
0.513s: Style/EmptyLines
0.505s: Style/StringLiterals
0.411s: Style/ExtraSpacing
0.380s: Style/ColonMethodCall
0.378s: Style/SpaceAroundOperators
0.375s: Style/SpaceInsideHashLiteralBraces
0.350s: Style/Semicolon
0.315s: Style/FormatString
0.306s: Style/IndentationConsistency
0.285s: Style/TrailingComma
0.232s: Style/StringLiteralsInInterpolation
````

After
````
Slowest Cops
2.437s: Style/ClosingParenthesisIndentation
2.260s: Style/MultilineOperationIndentation
1.184s: Style/IndentationWidth
1.079s: Lint/BlockAlignment
0.963s: Style/FirstParameterIndentation
0.776s: Lint/DeprecatedClassMethods
0.672s: Metrics/LineLength
0.601s: Style/SignalException
0.513s: Style/EmptyLines
0.471s: Style/StringLiterals
0.420s: Style/ExtraSpacing
0.388s: Style/SpaceAroundOperators
0.378s: Lint/Debugger
0.367s: Style/ColonMethodCall
0.365s: Style/SpaceInsideHashLiteralBraces
0.338s: Style/Semicolon
0.316s: Style/FormatString
0.297s: Style/IndentationConsistency
0.290s: Style/TrailingComma
0.215s: Style/StringLiteralsInInterpolation
````

@bbatsov I have the beginnings of this output in a branch, but I'm not sure whether this is something that you'd want in master: https://github.com/bbatsov/rubocop/compare/master...wli:performance